rest/pyproject.toml
[tool.poetry]
name = "kuber-rest-api"
version = "0.0.1"
description = "REST API for kuber"
authors = ["deepakdinesh1123 <d.deepakdinesh13@gmail.com>"]
readme = "README.md"
packages = [{include = "kuber_rest_api"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2"
django-rest-framework = "^0.1.0"
psycopg2 = "^2.9.6"
djangorestframework-simplejwt = "^5.2.2"
pre-commit = "^3.2.2"
django-cors-headers = "^3.14.0"
grpcio = "^1.54.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
